LayoutTests/fast/history/page-cache-clearing-expected.txt

Tests that clearing the page cache doesn't leave pages in the cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS window.internals.pageCacheSize() is 1
window.internals.clearPageCache()
PASS window.internals.pageCacheSize() is 0
PASS successfullyParsed is true

TEST COMPLETE
